+++ app/src/main/java/com/marcomadalin/olympus/domain/model/Set.kt
@@ -4,8 +4,8 @@ import com.marcomadalin.olympus.data.database.entities.SetEntity
 import com.marcomadalin.olympus.domain.model.enums.SetType
 
 data class Set(
-    var id : Int = 0,
-    var exerciseId : Int = 0,
+    var id : Long = 0,
+    var exerciseId : Long = 0,
     var weight : Int = 0,
     var reps : Int = 0,
     var rir : Int = 0,
